Fix bug of the last commit
In partial hessian sum, instead of using Eigen Map and Eigen copy, I tried to copy outer ptr, inner ptr, and nonzero values directly from gpu to cpu, but the program got segfault for some cases. The reason is that in full_to_reduced_matrix, when the matrix does not need to reduce, it use "reduced=full" to copy matrix. But with the modification in partial hessian sum, this line of code not copy nonzero values. I fixed it by adding Eigen Map before copying. I tested on walking elephant, which did not work, and it gets correct result now.